Time: 27 ms (27.37%), Space: 58.4 MB (5.61%) - LeetHub

+++ 1354-construct-target-array-with-multiple-sums/1354-construct-target-array-with-multiple-sums.java
@@ -19,7 +19,7 @@ class Solution {
             if(oldTop<=rem){
                 return false;
             }
-            Long newTop=oldTop-(rem*((int)(oldTop/rem)));
+            Long newTop=oldTop-(rem*(oldTop/rem));
             
             
             if(newTop==0){
